Fixed selection of screen number in video_stream_player.

+++ include/sdl_render_window.hh
@@ -13,7 +13,7 @@ public:
 		  const std::string &font_file,
 		  const std::string &home_dir_icon,
 		  const std::string &north_arrow_icon,
-		  uint8_t screen,
+		  uint16_t screen,
 		  bool fullstcreen);
 
   ~SDLRenderWindow();
@@ -29,7 +29,7 @@ private:
   std::string m_home_dir_icon;
   std::string m_north_arrow_icon;
   std::shared_ptr<SDLOSD> m_osd;
-  uint8_t m_screen;
+  uint16_t m_screen;
   SDL_Window *m_win;
   SDL_Renderer *m_renderer;
   SDL_Texture *m_texture;
