+++ src/video_stream_player.cc
@@ -32,7 +32,7 @@ int main(int argc, char* argv[]) {
   uint32_t packet_size;
   bool use_udp;
   bool fullscreen;
-  uint8_t screen;
+  uint16_t screen;
   std::string url;
   std::string font_file;
   std::string home_dir_icon;
@@ -46,7 +46,7 @@ int main(int argc, char* argv[]) {
      "the size of the packet buffer (the maximum size of a packet)")
     ("use_udp,U", po::bool_switch(&use_udp), "use the UDP protocol rather than TCP")
     ("fullscreen,f", po::bool_switch(&fullscreen), "make the render window full screen")
-    ("screen,s", po::value<uint8_t>(&screen), "the screen to display the video on")
+    ("screen,s", po::value<uint16_t>(&screen), "the screen to display the video on")
     ("url,u", po::value<std::string>(&url), "read from the specified URL")
     ("font", po::value<std::string>(&font_file), "the path to the OSD font file")
     ("home_dir_icon", po::value<std::string>(&home_dir_icon),
